# docs/swagger.yaml
basePath: /
definitions:
  app.Response:
    properties:
      data:
        type: object
      message:
        type: string
    type: object
  feiralivre.FeiraLivre:
    properties:
      areap:
        type: string
      bairro:
        type: string
      coddist:
        type: string
      codsubpref:
        type: string
      distrito:
        type: string
      id:
        type: integer
      lat:
        type: string
      lng:
        type: string
      logradouro:
        type: string
      nome:
        type: string
      numero:
        type: string
      referencia:
        type: string
      regiao5:
        type: string
      regiao8:
        type: string
      registro:
        type: string
      setcens:
        type: string
      subprefe:
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: burredis@gmail.com
  description: '"A RESTful API made with Golang and SQLite"'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: unico
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - application/json
      description: Check the api health
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
      tags:
      - API
  /feiraslivres:
    get:
      consumes:
      - application/json
      description: Search item by query
      parameters:
      - description: Search by distrito or regiao5 or nome or bairro
        in: query
        name: q
        type: string
      tags:
      - API
    post:
      consumes:
      - application/json
      description: Create a new item
      parameters:
      - description: ' '
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/feiralivre.FeiraLivre'
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      tags:
      - API
  /feiraslivres/:id:
    delete:
      consumes:
      - application/json
      description: Remove item by id
      responses:
        "202":
          description: ""
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      tags:
      - API
    get:
      consumes:
      - application/json
      description: Get item by id
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.Response'
      tags:
      - API
    patch:
      consumes:
      - application/json
      description: Update item by id
      parameters:
      - description: ' '
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/feiralivre.FeiraLivre'
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/app.Response'
      tags:
      - API
swagger: "2.0"